rename VChart -> Chart
