+++ src/views/landing/IntakeHero.vue
@@ -1,9 +1,8 @@
 <template>
-  <!-- overflow: hidden because the VChart autoresize isn't terribly precise and rounds up -->
   <div
     class="py-16 lg:py-64 flex flex-col items-center relative overflow-hidden"
   >
-    <VChart :option="chartOption" class="bg-chart" autoresize />
+    <Chart :option="chartOption" class="bg-chart" autoresize />
     <h1 class="font-bold text-white text-4xl lg:text-6xl mb-5 mx-4 text-center">
       Investigate
       <span class="text-green-400 uppercase">your NEAR</span> account
@@ -76,7 +75,7 @@ import { useSingle } from '@/composables/useSingle';
 import { Network } from '@/services/near/indexer/networks';
 import { Timeframe } from '@/services/timeframe';
 import { defineComponent, ref } from 'vue';
-import VChart from '@/components/VChart.vue';
+import Chart from '@/components/Chart.vue';
 import { useRouter } from 'vue-router';
 
 export default defineComponent({
@@ -85,7 +84,7 @@ export default defineComponent({
     TimeframeInput,
     NetworkInput,
     PrimaryButton,
-    VChart,
+    Chart,
   },
   name: 'intake-hero',
   setup() {
