+++ src/views/overview/widgets/TransactionRateWidget.vue
@@ -9,7 +9,7 @@
     </template>
     <template #default>
       <div class="flex flex-1" style="min-height: 250px">
-        <VChart
+        <Chart
           class="chart flex-shrink"
           :option="weeklyTransactionRateOption"
           autoresize
@@ -25,13 +25,13 @@ import { useNear } from '@/composables/useNear';
 import { useRecentActions } from '@/composables/useRecentActions';
 import { Timeframe } from '@/services/timeframe';
 import { defineComponent, ref, watch } from 'vue';
-import VChart from '@/components/VChart.vue';
+import Chart from '@/components/Chart.vue';
 import DashboardCard from '../DashboardCard.vue';
 
 export default defineComponent({
   components: {
     DashboardCard,
-    VChart,
+    Chart,
   },
   setup() {
     const { account, network } = useNear();
